api-w-caching/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
orjson>=3.10
psycopg2-binary>=2.9
pydantic>=2.0
python-dotenv>=1.0
